fixed test issue (moved function so it doesn't import state)
