+++ app/src/state/actions/util-actions.ts
@@ -1,13 +1,7 @@
 import { action } from "mobx";
 import { ErrorCodes } from "../../util/error-codes";
 import { pushError, pushSnackbar } from "./general-actions";
-import {
-  ColumnInfo,
-  DisplayedColumnTypes,
-  assertEventSettings,
-  assertPointSettings,
-  isServerResponseError
-} from "@tsconline/shared";
+import { ColumnInfo, assertEventSettings, assertPointSettings, isServerResponseError } from "@tsconline/shared";
 import { state } from "../state";
 /**
  * Since we hash by name only to allow consistency between facies maps and
@@ -121,15 +115,4 @@ export function checkIfDccDataIsInRange(dccColumn: ColumnInfo, userTopAge: numbe
     dccColumn = refCol;
   }
   return true;
-}
-
-export function attachTscPrefixToName(name: string, displayType: DisplayedColumnTypes): string {
-  switch (displayType) {
-    case "RootColumn":
-    case "MetaColumn":
-    case "BlockSeriesMetaColumn":
-      return `class datastore.${displayType}:` + name;
-    default:
-      return `class datastore.${displayType}Column:` + name;
-  }
 }
